+++ docs/changelog.rst
@@ -6,6 +6,8 @@ Changelog
 0.12.0 - TBD
    * Removed support for python3.6
    * Changed packaging to use hatch
+   * Pip is made to be greater than 23 when a venv is used
+     unless ``VENVSTARTER_UPGRADE_PIP=0``
 
 .. _release-0.11.0: